packages/control-plane/requirements.txt
fastapi>=0.111.0,<1
greenlet>=3.0.0
uvicorn[standard]>=0.29.0,<1
sqlalchemy>=2.0.0,<3
aiosqlite>=0.20.0,<1
asyncpg>=0.29.0,<1
python-jose[cryptography]>=3.3.0,<4
pydantic>=2.7.0,<3
kubernetes-asyncio>=29.0.0,<31
